% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_residuals.R
\name{calc_residuals}
\alias{calc_residuals}
\title{Calculate residuals}
\usage{
calc_residuals()
}
\value{
A data frame of residuals
}
\description{
Calculate residuals
}
\examples{
\dontrun{
calc_residuals()
}
}
\concept{helpers}
