% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_least_squares.R
\name{calc_sum_squares}
\alias{calc_sum_squares}
\title{Calculate least squares}
\usage{
calc_sum_squares()
}
\value{
Numeric value of the sum of the squares of the
}
\description{
Calculate least squares
}
\examples{
\dontrun{
calc_sum_squares()
}
}
