% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_rm.R
\name{cg_rm}
\alias{cg_rm}
\title{Apply the Rasch model to dichotomous data.}
\usage{
cg_rm(sensitivity)
}
\value{
A list of outputs
}
\description{
Apply the Rasch model to dichotomous data.
}
\examples{
\dontrun{
rm()
}
}
