% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wright_map.R
\name{wright_map}
\alias{wright_map}
\title{Plot a wright map of a Rasch model}
\usage{
wright_map(mod)
}
\arguments{
\item{mod}{A Rasch model created by \code{cg_rm()}}
}
\value{
A ggplot object.
}
\description{
Plot a wright map of a Rasch model
}
\examples{
mod <- dichotomous |>
cg_rm()

mod |> wright_map()

}
