% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_rm.R
\name{cg_rm}
\alias{cg_rm}
\title{Apply the Rasch model to dichotomous data.}
\usage{
cg_rm(df, sensitivity)
}
\arguments{
\item{df}{A dataframe of candidate response data to dichotomous items. The
dataframe should be in long format with person, item and response variables.
Missing data is allowed.}
}
\value{
A list of outputs
}
\description{
Apply the Rasch model to dichotomous data.
}
\examples{
data(dichotomous)
model <- dichotomous |>
  cg_rm()
model
}
